# openapi.yaml
openapi: 3.0.1
info:
  title: SurvHey BackEnd
  description: 'This is the OAS for the BackEnd of the SurvHey Project.'
  version: 1.0.0
servers:
  - url: http://localhost:8080/v2
...
paths:
  /user:
    post:
      tags:
        - user
      summary: Create user
      description: Registers a new user
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        200:
          description: successful operation
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    get:
      tags:
        - user
      summary: Get user by api_key
      description: This can only be done by the logged-in user and will return a user object
      operationId: getUser
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - ApiKeyAuth: []
    delete:
      tags:
        - user
      summary: Delete user
      description: This can only be done by the logged-in user
      operationId: deleteUser
      responses:
        200:
          description: successful operation
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - ApiKeyAuth: []
  /user/login:
    post:
      tags:
        - user
      summary: Logs user into the system
      operationId: loginUser
      requestBody:
        description: Object that stores user login data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthKey'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /user/logout:
    get:
      tags:
        - user
      summary: Logs out current logged in user session
      description: This can only be done by the logged-in user
      operationId: logoutUser
      responses:
        200:
          description: successful operation
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - ApiKeyAuth: []
...
  /surveys:
    get:
      tags:
        - survey
      summary: Return all surveys for the logged in user
      description: All surveys for the corresponding user are returned
      operationId: getSurveys
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Survey'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - ApiKeyAuth: []
...
components:
  schemas:
    User:
      required:
        - id
        - email
        - password
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
    UserLogin:
      required:
        - id
        - email
        - password
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        password:
          type: string
    Survey:
      required:
        - id
        - name
        - user
        - questions
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        user:
          $ref: '#/components/schemas/User'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
...
    Question:
      required:
        - id
        - survey_id
        - text
      type: object
      properties:
        id:
          type: integer
          format: int64
        survey_id:
          type: integer
          format: int64
        text:
          type: string
        mode:
          type: string
          enum: [radio, check, text, rating, dropdown, nps, consent]
        answerOptions:
          type: array
          items:
            $ref: '#/components/schemas/AnswerOption'
    AnswerOption:
      required:
        - id
        - question_id
        - content
      type: object
      properties:
        id:
          type: integer
          format: int64
        question_id:
          type: integer
          format: int64
        content:
          oneOf:
            - type: string
            - type: number
            - type: integer
            - type: boolean
...
    AuthKey:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        authKey:
          type: string
        expiry:
          type: string
          format: date-time
    ApiError:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: api_key
      in: header